requirements.txt
appdirs==1.4.3
astroid==2.3.2
autopep8==1.4.4
box2d-py==2.3.8
cloudpickle==1.2.2
colorama==0.4.1
future==0.18.2
gym==0.15.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.3
opencv-python==4.1.1.26
Pillow==6.2.1
pycodestyle==2.5.0
pygame==1.9.6
pyglet==1.3.2
pygubu==0.9.8.4
pylint==2.4.3
scipy==1.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
